% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterModel.R
\name{betterModel}
\alias{betterModel}
\title{betterModel_ helper function}
\usage{
betterModel(statstable_pre, statstable_post_hyperbolic = NULL,
  statstable_post_cubic = NULL, selection_method = "SSE")
}
\arguments{
\item{statstable_pre}{A data.table object, containing the output of \code{statisticsList_()} of the
calculated regression parameters (form the provided calibration data).}

\item{statstable_post_hyperbolic}{A data.table object, containing the output of \code{statisticsList_()}
of the calculated regression parameters form the calibration data corrected with hyperbolic regression.}

\item{statstable_post_cubic}{A data.table object, containing the output of \code{statisticsList_()}
of the calculated regression parameters form the calibration data corrected with cubic regression.}

\item{selection_method}{A character string. The method used to select the regression algorithm to correct the
respective CpG site. This is by default the sum of squared errors ("SSE"). The second option is "RelError",
which selects the regression method based on the theoretical relative error after correction. This metric is
calculated by correcting the calibration data with both the hyperbolic regression and the cubic regression and
using them again as input data to calculate the 'goodness of fit'-metrics.}
}
\description{
Internal function to select the better model between hyperbolic regression and
  cubic regression.
}
