% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_plot.R
\name{calibration_plot}
\alias{calibration_plot}
\title{calibration_plot helper function}
\usage{
calibration_plot(plotlist, coef_hyper, coef_cubic, plot_textsize, minmax)
}
\arguments{
\item{plotlist}{A ggplot object containing a regression plot without
regression curves (output of \code{regression_utility()}).}

\item{coef_hyper}{A list containing the regression parameters of
the hyperbolic regression equation.}

\item{coef_cubic}{A list containing the regression parameters of
the cubic regression equation.}

\item{plot_textsize}{A integer value. The textsize of the
resulting plots (default: 16).}

\item{minmax}{A logical, indicating which equations are used for
BiasCorrection (default: FALSE). If TRUE, equations are used that include
the respective minima and maxima of the provided data.}
}
\value{
The function returns a list containing calibration plots.
}
\description{
Internal function to carry out the plotting of the
  calibrations curves.
}
