% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_up}
\alias{clean_up}
\title{clean_up helper function}
\usage{
clean_up(plotdir, csvdir)
}
\arguments{
\item{plotdir}{A character string. Path to the folder,
where plots are saved.}

\item{csvdir}{A character string. Path to the folder,
where resulting tables are saved.}
}
\value{
This function silently cleans up the current session and removes
both, the \code{plotdir}- and the \code{csvdir}-folders. It furthermore resets
the \code{future}-backend to plan = "sequential".
}
\description{
Internal function to clean up directories.
}
\examples{
plotdir <- file.path(tempdir(), "plots/")
csvdir <- file.path(tempdir(), "csv/")

clean_up(plotdir, csvdir)

}
\seealso{
\link[future]{plan}
}
