% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanDT.R
\name{cleanDT_}
\alias{cleanDT_}
\title{cleanDT helper function}
\usage{
cleanDT_(datatable, description, type, logfilename)
}
\description{
Function, that checks the formatting of imported files and prepares them to
  be applicable for the following pcr-bias correction steps.
}
