% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanUp_}
\alias{cleanUp_}
\title{cleanUp_ helper function}
\usage{
cleanUp_(plotdir, csvdir)
}
\arguments{
\item{plotdir}{A character string. Path to the folder, where plots are saved.}

\item{csvdir}{A character string. Path to the folder, where resulting tables are saved.}
}
\description{
Internal function to clean up directories.
}
