% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{createBarErrorPlots_}
\alias{createBarErrorPlots_}
\title{createBarErrorPlots helper function}
\usage{
createBarErrorPlots_(statstable_pre, statstable_post, rv, type, b = NULL,
  headless = FALSE, plotdir, logfilename, mode = NULL)
}
\description{
Function to create relative-error bar plots.
}
