% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biascorrection.R
\docType{data}
\name{example._plot.df_agg}
\alias{example._plot.df_agg}
\title{example._plot.df_agg}
\description{
A data.table containing the aggregated
  calibration data for CpG site 1
  to create an example plot.
}
\keyword{data}
