% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biascorrection.R
\docType{data}
\name{example._plot_coef_c}
\alias{example._plot_coef_c}
\title{example._plot_coef_c}
\description{
A list containing exemplary coefficients
  of the cubic regression equation for CpG site 1
  to create an example plot.
}
\keyword{data}
