% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biascorrection.R
\docType{data}
\name{example._plot_coef_h}
\alias{example._plot_coef_h}
\title{example._plot_coef_h}
\description{
A list containing exemplary coefficients
of the hyperbolic regression equation for CpG site 1
to create an example plot.
}
\keyword{data}
