% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasCorrection.R
\docType{data}
\name{example.data_calibration}
\alias{example.data_calibration}
\title{example.data_calibration}
\description{
A list containing the calibration data ($dat) and the colnames ($vec_cal) needed by the algorithms to perform the bias correction.
}
\keyword{data}
