% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biascorrection.R
\docType{data}
\name{example.data_experimental}
\alias{example.data_experimental}
\title{example.data_experimental}
\description{
A list containing the experimental data ($dat) and the
  colnames ($vec_cal) needed by the algorithms to perform the bias
  correction.
}
\keyword{data}
