% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_timestamp}
\alias{get_timestamp}
\title{get_timestamp helper function}
\usage{
get_timestamp()
}
\value{
This function takes no argument and returns a formatted timestamp
  of the current system time, which can be integrated e.g. into a filename.
}
\description{
Internal function to get the current timestamp to
  write it to filenames.
}
\examples{
get_timestamp()

}
\seealso{
\link{Sys.time}
}
