% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{onStart_}
\alias{onStart_}
\title{onStart_ helper function}
\usage{
onStart_(plotdir, csvdir, logfilename)
}
\arguments{
\item{plotdir}{A character string. Path to the folder, where plots are saved.}

\item{csvdir}{A character string. Path to the folder, where resulting tables are saved.}

\item{logfilename}{A character string. Path to the logfile to save the log messages.}
}
\description{
Internal function, that initializes plotdir, csvdir and logfilename.
}
