% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plottingUtility_}
\alias{plottingUtility_}
\title{plottingUtility helper function}
\usage{
plottingUtility_(data, plotlistR, type, samplelocusname, locus_id = NULL,
  rv, mode = NULL, headless = FALSE, plotdir, logfilename,
  minmax = FALSE)
}
\description{
Function to carry out the plotting of the calibrations curves.
}
