% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regressionUtility_}
\alias{regressionUtility_}
\title{regressionUtility helper function}
\usage{
regressionUtility_(data, samplelocusname, locus_id = NULL, rv,
  mode = NULL, headless = FALSE)
}
\description{
Function to carry out the regression calculations.
}
