% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rsq}
\alias{rsq}
\title{R-squared helper function}
\usage{
rsq(true, fitted)
}
\description{
Caclulates the Coefficient of determinition (R-squared,
  \url{https://en.wikipedia.org/wiki/Coefficient_of_determination}).
}
