% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_utils.R
\name{solvingEquations_}
\alias{solvingEquations_}
\title{solvingEquations helper function}
\usage{
solvingEquations_(datatable, regmethod, type, rv, mode = NULL,
  logfilename)
}
\description{
Solving hyperbolic and cubic regression.
}
