% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticsList.R
\name{statisticsList_}
\alias{statisticsList_}
\title{statisticsList_ helper function}
\usage{
statisticsList_(resultlist, minmax = FALSE)
}
\arguments{
\item{resultlist}{A list object. The results_list output of \code{regressionUtility_()}.}

\item{minmax}{A logical, indicating which equations are used for BiasCorrection (default: FALSE). If TRUE, equations are used
that include the respective minima and maxima of the provided data.}
}
\description{
Internal function that converts the results_list (output of \code{regressionUtility_()})
  into a data.table object.
}
