% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{substitutions_create}
\alias{substitutions_create}
\title{substitutions_create helper function}
\usage{
substitutions_create()
}
\value{
This function takes no argument and initializes an empty
  `data.table` to hold the substituted values, if substitutions occur during
  BiasCorrection.
}
\description{
Internal function to initialize a data.table object
  to store the substitutions.
}
\examples{
substitutions <- substitutions_create()
class(substitutions)

}
