% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type2files.R
\name{type2_filereq}
\alias{type2_filereq}
\title{type2_filereq helper function}
\usage{
type2_filereq(filelist, rv, logfilename)
}
\arguments{
\item{filelist}{A list object, containing all type 2 data
calibrations files.}

\item{rv}{A list object. A list that contains additional objects needed
for the algorithms.}

\item{logfilename}{A character string. Path to a file to save the log
messages. Default = "./log.txt"}
}
\description{
Internal function to check type 2 data file requirements.
}
