% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_log}
\alias{write_log}
\title{write_log helper function}
\usage{
write_log(message, logfilename)
}
\arguments{
\item{message}{A character string containing the log message.}

\item{logfilename}{A character string. Path to a file to save the log
messages (default = file.path(tempdir(), "log.txt")).}
}
\value{
The function prints the logging message to the console and writes it
to the local logfile, specified with \code{logfilename}.
}
\description{
Internal function to write log-messages to the file
specified in logfilename.
}
\examples{
message <- "This is a logmessage"
logfilename <- paste0(tempdir(), "/log.txt")

write_log(message, logfilename)

}
