% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeCSV_}
\alias{writeCSV_}
\title{writeCSV_ helper function}
\usage{
writeCSV_(table, filename)
}
\arguments{
\item{table}{A data.table object to store on the local file system}

\item{filename}{The file name (including the path) to store \code{table}.}
}
\description{
Internal function to store the created tables in csv files.
}
