% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasCorrection.R
\name{BiasCorrection}
\alias{BiasCorrection}
\title{Correct PCR Bias in Quantitative DNA Methylation Analyses.}
\usage{
BiasCorrection(experimental, calibration, samplelocusname, type = 1,
  csvdir = "./csvdir", plotdir = "./plotdir",
  logfilename = "./log.txt")
}
\arguments{
\item{experimental}{A character string. Path to the file containing the raw methylation
values of the samples under investigation.}

\item{calibration}{A character string. In type 1 data (one locus in many samples, e.g. pyrosequencing data):
Path to the file containing the raw methylation values of the calibration samples.
In type 2 data (many loci in one sample, e.g. next-generation sequencing data or microarray data):
Path to the folder that contains at least 4 calibration files (one file per calibration step).
Please refere to the FAQ for more detailed information on the specific file requirements
(\url{https://raw.githubusercontent.com/kapsner/BiasCorrector/master/FAQ.md}).}

\item{samplelocusname}{A character string. In type 1 data: locus name - name of the gene locus under investigation.
In type 2 data: sample name - name of the sample under investigation.}

\item{type}{A single integer. Type of data to be corrected: either "1" (one locus in many samples, e.g. pyrosequencing data)
or "2" (many loci in one sample, e.g. next-generation sequencing data or microarray data).}

\item{csvdir}{A character string. Path to a folder to store the resulting tables. Default = "./csvdir".}

\item{plotdir}{A character string. Path to a folder to store the resulting plots. Default = "./plotdir".}

\item{logfilename}{A character string. Path to a file to save the log messages. Default = "./log.txt"}
}
\value{
TRUE, if the correction of PCR measurment biases succeeds. If the correction fails, an error message
  is returned.
}
\description{
This function implements the algorithms described by Moskalev et. al in their
 article 'Correction of PCR-bias in quantitative DNA methylation studies by means of cubic polynomial regression',
 published 2011 in Nucleic acids research, Oxford University Press (\url{https://doi.org/10.1093/nar/gkr213}).
}
