% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{createbarerrorplots}
\alias{createbarerrorplots}
\title{createbarerrorplots helper function}
\usage{
createbarerrorplots(statstable_pre, statstable_post, rv, type,
  locus_id = NULL, headless = FALSE, plotdir, logfilename,
  mode = NULL)
}
\arguments{
\item{statstable_pre}{A data.table object, containing the output of
\code{statisticsList_()} of the calculated regression parameters (form
the provided calibration data).}

\item{statstable_post}{A data.table object, containing the output of
\code{statisticsList_()} of the calculated regression parameters form
the corrected calibration data.}

\item{rv}{A list object. A list that contains additional objects needed
for the algorithms.}

\item{type}{A single integer. Type of data to be corrected: either "1"
(one locus in many samples, e.g. pyrosequencing data) or "2" (many loci
in one sample, e.g. next-generation sequencing data or microarray data).}

\item{locus_id}{A character string. Default: NULL. ID of the respective
locus (only used in type 2 correction).}

\item{headless}{A logical (default: FALSE). Indicates, if the function is
called from within a shiny app (default) or just from the console without
a graphical user interface.}

\item{plotdir}{A character string. Path to the folder,
where plots are saved.}

\item{logfilename}{A character string. Path to a file to save the log
messages. Default = "./log.txt"}

\item{mode}{A character string. Default: NULL. Used to indicate "corrected"
calibration data.}
}
\description{
Internal function to create relative-error bar plots.
}
