% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_text_input}
\alias{handle_text_input}
\title{handle_text_input helper function}
\usage{
handle_text_input(textinput)
}
\arguments{
\item{textinput}{A character string with the textinput to perform these
predefined regular expressions on.}
}
\description{
Internal function to remove punctuation and unneeded stuff
  from user inputs with regular expressions.
}
