% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{on_start}
\alias{on_start}
\title{on_start helper function}
\usage{
on_start(plotdir, csvdir, logfilename, parallel)
}
\arguments{
\item{plotdir}{A character string. Path to the folder,
where plots are saved.}

\item{csvdir}{A character string. Path to the folder,
where resulting tables are saved.}

\item{logfilename}{A character string. Path to the logfile to save the log
messages.}

\item{parallel}{A boolean. If TRUE (default), initializing
`future::plan("multiprocess")` before running the code.}
}
\value{
This function silently creates the directories`plotdir` and `csvdir`
  on the local filesystem and initializes the logfile, specified with
  `logfilename`. Furthermore, if `parallel = TRUE`, the `future`-backend is
  initialized.
}
\description{
Internal function, that initializes plotdir,
  csvdir and logfilename.
}
\examples{
plotdir <- paste0(tempdir(), "/plots/")
csvdir <- paste0(tempdir(), "/csv/")
logfilename <- "log.txt"
parallel <- TRUE

on_start(plotdir, csvdir, logfilename, parallel)

}
\seealso{
\link[future]{plan}
}
