% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regression_utility}
\alias{regression_utility}
\title{regression_utility helper function}
\usage{
regression_utility(data, samplelocusname, locus_id = NULL, rv,
  mode = NULL, headless = FALSE, logfilename, minmax)
}
\arguments{
\item{data}{A data.table object that contains the calibration data.}

\item{samplelocusname}{A character string. In type 1 data: locus name -
name of the gene locus under investigation. In type 2 data: sample name -
name of the sample under investigation.}

\item{locus_id}{A character string. Default: NULL. ID of the respective
locus (only used in type 2 correction).}

\item{rv}{A list object. A list that contains additional objects needed
for the algorithms.}

\item{mode}{A character string. Default: NULL. Used to indicate "corrected"
calibration data.}

\item{headless}{A logical (default: FALSE). Indicates, if the function is
called from within a shiny app (default) or just from the console without
a graphical user interface.}

\item{logfilename}{A character string. Path to a file to save the log
messages. Default = "./log.txt"}

\item{minmax}{A logical, indicating which equations are used for
BiasCorrection (default: FALSE). If TRUE, equations are used that include
the respective minima and maxima of the provided data.}
}
\description{
Internal function to carry out the regression calculations.
}
