% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{substitutions_create}
\alias{substitutions_create}
\title{substitutions_create helper function}
\usage{
substitutions_create()
}
\description{
Internal function to initialize a data.table object
  to store the substitutions.
}
