% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{substitutionsCreate_}
\alias{substitutionsCreate_}
\title{substitutionsCreate helper function}
\usage{
substitutionsCreate_()
}
\description{
Function to initialize a data.table object to store the substitutions.
}
