% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_csv}
\alias{write_csv}
\title{write_csv helper function}
\usage{
write_csv(table, filename)
}
\arguments{
\item{table}{A data.table object to store on the local file system}

\item{filename}{The file name (including the path) to store \code{table}.}
}
\description{
Internal function to store the created tables in csv files.
}
